// project_name/src/modules/category/dto/update-category.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreateCategoryDto } from './create-category.dto';
import { IsNotEmpty } from 'class-validator';

export class UpdateCategoryDto extends PartialType(CreateCategoryDto) {
  @IsNotEmpty()
  name: string;
  @IsNotEmpty()
  slug: string;
  description: string;
  hot: number | 0;
  status: number;
}